+++ front/migrations/db/migration_199.sql
@@ -1,8 +1,8 @@
 -- Migration created on Feb 02, 2025
-ALTER TABLE "public"."conversations" ADD COLUMN "currentThreadVersion" BIGINT NOT NULL DEFAULT 0;
+ALTER TABLE "public"."conversations" ADD COLUMN "currentThreadVersion" INTEGER NOT NULL DEFAULT 0;
 ALTER TABLE "public"."messages" ADD COLUMN "nextVersionMessageId" BIGINT;
 ALTER TABLE "public"."messages" ADD COLUMN "previousVersionMessageId" BIGINT;
-ALTER TABLE "public"."messages" ADD COLUMN "threadVersions" BIGINT[] DEFAULT ARRAY[0]::BIGINT[];
+ALTER TABLE "public"."messages" ADD COLUMN "threadVersions" INTEGER[] DEFAULT ARRAY[0]::INTEGER[];
 CREATE UNIQUE INDEX CONCURRENTLY "messages_conversation_id_rank_version_parent_id" ON "messages" ("conversationId", "rank", "version", "parentId");
 CREATE UNIQUE INDEX CONCURRENTLY "messages_conversation_id_rank_version_thread_version" ON "messages" ("conversationId", "rank", "version", ("threadVersions"[1]));
 DROP INDEX IF EXISTS "messages_conversation_id_rank_version";
